requirements.txt
PyQt5
opencv-python
mediapipe
numpy
absl-py
